refactor: move icon.svg to icons folder
